[Feat] #39 - DiaryScreen API 연결 완료
