+++ lib/presentaion/diary/view/diary_screen.dart
@@ -1,14 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:money_planet/presentaion/diary/view/diary_calendar_screen.dart';
-import 'package:money_planet/presentaion/diary/view/diary_list_screen.dart';
-import 'package:money_planet/presentaion/diary/view/empty_diary_screen.dart';
 import 'package:money_planet/presentaion/diary/viewModel/diary_viewModel.dart';
+import 'package:money_planet/presentaion/home/view/home_third_section.dart';
 
 import '../../../global/components/floating_dial_button.dart';
 import '../../../global/theme/colors.dart';
 import '../../../global/theme/textStyles.dart';
+import '../../home/viewModel/home_viewModel.dart';
 import '../../myPage/view/hint_screen.dart';
-import '../model/diary_model.dart';
 
 class DiaryScreen extends StatefulWidget {
   const DiaryScreen({super.key});
@@ -18,21 +16,11 @@ class DiaryScreen extends StatefulWidget {
 }
 
 class _DiaryScreenState extends State<DiaryScreen> {
-  var viewModel = DiaryViewModel();
+  final DiaryViewModel viewModel = DiaryViewModel();
+  final HomeViewModel homeViewModel = HomeViewModel();
 
   @override
   Widget build(BuildContext context) {
-    int selectedMonth = viewModel.selectedMonth;
-
-    // 필터링된 데이터만 가져오기
-    final filteredData =
-    DiaryMockData.where(
-          (item) => item.date.month == viewModel.selectedMonth,
-    ).toList();
-
-    // 날짜별로 그룹핑된 Map 생성
-    final groupedData = viewModel.groupByDate(filteredData);
-
     return Scaffold(
       backgroundColor: neutral_900,
       floatingActionButton: floatingDialButton(context),
@@ -85,218 +73,11 @@ class _DiaryScreenState extends State<DiaryScreen> {
               SizedBox(height: 29),
 
               // 가계부
-              Container(
-                decoration: BoxDecoration(
-                  borderRadius: BorderRadius.only(
-                    topLeft: Radius.circular(40),
-                    topRight: Radius.circular(40),
-                  ),
-                  color: Colors.white,
-                ),
-                child: ConstrainedBox(
-                  constraints: BoxConstraints(
-                    minHeight: MediaQuery.of(context).size.height - 100, // 화면의 높이에서 헤더 등을 제외한 값
-                  ),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.start,
-                    children: [
-                      // 달 바꾸는 부분 & 달력 아이콘
-                      Padding(
-                        padding: const EdgeInsets.symmetric(
-                          horizontal: 20,
-                          vertical: 30,
-                        ),
-                        child: Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          children: [
-                            Row(
-                              children: [
-                                IconButton(
-                                  onPressed: () => changeMonth(-1),
-                                  icon: const Icon(Icons.chevron_left),
-                                ),
-                                Text(
-                                  "$selectedMonth 월",
-                                  style: customTextStyle(
-                                    fontFamily: Pretendard_Semibold_24,
-                                    color: neutral_1100,
-                                  ),
-                                ),
-                                IconButton(
-                                  onPressed: () => changeMonth(1),
-                                  icon: const Icon(Icons.chevron_right),
-                                ),
-                              ],
-                            ),
-
-                            Padding(
-                              padding: EdgeInsets.only(right: 10),
-                              child: Container(
-                                decoration: BoxDecoration(
-                                  borderRadius: BorderRadius.circular(26),
-                                  color: primary_200,
-                                ),
-                                width: 48,
-                                height: 48,
-                                child: IconButton(
-                                  onPressed:
-                                      () => Navigator.push(
-                                    context,
-                                    MaterialPageRoute(
-                                      builder: (_) => DiaryCalendarScreen(),
-                                    ),
-                                  ),
-                                  iconSize: 24,
-                                  icon: Icon(
-                                    Icons.calendar_today,
-                                    color: Colors.white,
-                                  ),
-                                ),
-                              ),
-                            ),
-                          ],
-                        ),
-                      ),
-
-                      // 지출, 수입
-                      Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 20),
-                        child: Column(
-                          children: [
-                            Row(
-                              children: [
-                                SizedBox(width: 10),
-                                Text(
-                                  "지출",
-                                  style: customTextStyle(
-                                    fontFamily: Pretendard_Semibold_16,
-                                    color: neutral_400,
-                                  ),
-                                ),
-                                Spacer(),
-                                Text(
-                                  "475,180 원",
-                                  style: customTextStyle(
-                                    fontFamily: Pretendard_Medium_24,
-                                    color: Colors.black,
-                                  ),
-                                ),
-                                SizedBox(width: 10),
-                              ],
-                            ),
-                            Row(
-                              children: [
-                                SizedBox(width: 10),
-                                Text(
-                                  "수입",
-                                  style: customTextStyle(
-                                    fontFamily: Pretendard_Semibold_16,
-                                    color: neutral_400,
-                                  ),
-                                ),
-                                Spacer(),
-                                Text(
-                                  "0 원",
-                                  style: customTextStyle(
-                                    fontFamily: Pretendard_Medium_24,
-                                    color: Colors.black,
-                                  ),
-                                ),
-                                SizedBox(width: 10),
-                              ],
-                            ),
-                          ],
-                        ),
-                      ),
-
-                      SizedBox(height: 25),
-
-                      // 구분선
-                      Container(
-                        height: 12,
-                        width: double.infinity,
-                        color: neutral_100,
-                      ),
-
-                      SizedBox(height: 35),
-
-                      // 리스트 보여주는 곳
-                      groupedData.isEmpty
-                          ? EmptyDiaryScreen()
-                          : Column(
-                        children: [
-                          // daily, weekly, monthly 선택 버튼
-                          Padding(
-                            padding: const EdgeInsets.symmetric(
-                              horizontal: 5,
-                            ),
-                            child: Row(
-                              mainAxisAlignment:
-                              MainAxisAlignment.spaceEvenly,
-                              children: [
-                                for (var viewType in [
-                                  'Daily',
-                                  'Weekly',
-                                  'Monthly',
-                                ])
-                                  ElevatedButton(
-                                    onPressed: () => changeView(viewType),
-                                    style: ElevatedButton.styleFrom(
-                                      backgroundColor:
-                                      viewModel.selectedView == viewType
-                                          ? primary_400
-                                          : primary_050,
-                                      foregroundColor:
-                                      viewModel.selectedView == viewType
-                                          ? Colors.white
-                                          : primary_400,
-                                      shape: RoundedRectangleBorder(
-                                        borderRadius: BorderRadius.circular(
-                                          6,
-                                        ),
-                                        side: const BorderSide(
-                                          color: Colors.blue,
-                                        ),
-                                      ),
-                                    ),
-                                    child: Text(
-                                      viewType,
-                                      style: customTextStyle(
-                                        fontFamily: Pretendard_Semibold_18,
-                                      ),
-                                    ),
-                                  ),
-                              ],
-                            ),
-                          ),
-                          const SizedBox(height: 30),
-                          DiaryListScreen(groupedData: groupedData),
-                        ],
-                      ),
-
-                      const SizedBox(height: 20),
-                    ],
-                  ),
-                ),
-              ),
+              HomeThirdSection(viewModel: homeViewModel),
             ],
           ),
         ),
       ),
     );
-  }
-
-  void changeView(String view) {
-    setState(() {
-      viewModel.selectedView = view;
-    });
-  }
-
-  void changeMonth(int delta) {
-    setState(() {
-      viewModel.selectedMonth += delta;
-      if (viewModel.selectedMonth < 1) viewModel.selectedMonth = 1;
-      if (viewModel.selectedMonth > 12) viewModel.selectedMonth = 12;
-    });
   }
 }
