+++ lib/presentaion/home/view/home_third_section.dart
@@ -5,6 +5,8 @@ import 'package:money_planet/global/theme/textStyles.dart';
 
 import '../../../global/components/daily_category_item.dart';
 import '../../../network/Daily/Response/DailyCategoryResponseDTO.dart';
+import '../../diary/view/diary_calendar_screen.dart';
+import '../../diary/view/empty_diary_screen.dart';
 import '../viewModel/home_viewModel.dart';
 
 class HomeThirdSection extends StatefulWidget {
@@ -104,173 +106,185 @@ class HomeThirdSectionState extends State<HomeThirdSection> {
     final totalIncome = _incomeStats.fold(0, (sum, e) => sum + e.amount);
     final totalExpense = _expenseStats.fold(0, (sum, e) => sum + e.amount);
 
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 23),
-      child: Container(
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(12),
-          color: Colors.white,
-        ),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            /// Month selector
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 20),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Row(
-                    children: [
-                      IconButton(
-                        onPressed: () => changeMonth(-1),
-                        icon: const Icon(Icons.chevron_left),
-                      ),
-                      Text(
-                        "$selectedMonth월",
-                        style: customTextStyle(
-                          fontFamily: Pretendard_Semibold_24,
-                          color: neutral_1100,
-                        ),
-                      ),
-                      IconButton(
-                        onPressed: () => changeMonth(1),
-                        icon: const Icon(Icons.chevron_right),
-                      ),
-                    ],
-                  ),
-                  TextButton(
-                    onPressed: () {},
-                    child: Text(
-                      "자세히보기",
+    return Container(
+      decoration: BoxDecoration(
+        borderRadius: BorderRadius.circular(12),
+        color: Colors.white,
+      ),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          /// Month selector
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 20),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Row(
+                  children: [
+                    IconButton(
+                      onPressed: () => changeMonth(-1),
+                      icon: const Icon(Icons.chevron_left),
+                    ),
+                    Text(
+                      "$selectedMonth월",
                       style: customTextStyle(
-                        fontFamily: Pretendard_Medium_12,
-                        color: neutral_300,
+                        fontFamily: Pretendard_Semibold_24,
+                        color: neutral_1100,
+                      ),
+                    ),
+                    IconButton(
+                      onPressed: () => changeMonth(1),
+                      icon: const Icon(Icons.chevron_right),
+                    ),
+                  ],
+                ),
+                Container(
+                  decoration: BoxDecoration(
+                    borderRadius: BorderRadius.circular(26),
+                    color: primary_200,
+                  ),
+                  width: 44,
+                  height: 44,
+                  child: IconButton(
+                    onPressed:
+                        () => Navigator.push(
+                      context,
+                      MaterialPageRoute(
+                        builder: (_) => DiaryCalendarScreen(),
+                      ),
+                    ),
+                    iconSize: 24,
+                    icon: Icon(
+                      Icons.calendar_today,
+                      color: Colors.white,
+                    ),
+                  ),
+                ),
+              ],
+            ),
+          ),
+
+          /// Expense
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 20),
+            child: Row(
+              children: [
+                const SizedBox(width: 10),
+                Text(
+                  "지출",
+                  style: customTextStyle(
+                    fontFamily: Pretendard_Semibold_16,
+                    color: neutral_400,
+                  ),
+                ),
+                const Spacer(),
+                Text(
+                  "${formatAmount(totalExpense)} 원",
+                  style: customTextStyle(
+                    fontFamily: Pretendard_Medium_24,
+                    color: Colors.black,
+                  ),
+                ),
+                const SizedBox(width: 10),
+              ],
+            ),
+          ),
+
+          /// Income
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 20),
+            child: Row(
+              children: [
+                const SizedBox(width: 10),
+                Text(
+                  "수입",
+                  style: customTextStyle(
+                    fontFamily: Pretendard_Semibold_16,
+                    color: neutral_400,
+                  ),
+                ),
+                const Spacer(),
+                Text(
+                  "${formatAmount(totalIncome)} 원",
+                  style: customTextStyle(
+                    fontFamily: Pretendard_Medium_24,
+                    color: Colors.black,
+                  ),
+                ),
+                const SizedBox(width: 10),
+              ],
+            ),
+          ),
+
+          const SizedBox(height: 20),
+
+          /// View Type Buttons
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 5),
+            child: Row(
+              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+              children: [
+                for (var viewType in ['Daily', 'Weekly', 'Monthly'])
+                  ElevatedButton(
+                    onPressed: () => changeView(viewType),
+                    style: ElevatedButton.styleFrom(
+                      backgroundColor:
+                          selectedView == viewType ? primary_400 : primary_050,
+                      foregroundColor:
+                          selectedView == viewType ? Colors.white : primary_400,
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(6),
+                        side: const BorderSide(color: Colors.blue),
+                      ),
+                    ),
+                    child: Text(
+                      viewType,
+                      style: customTextStyle(
+                        fontFamily: Pretendard_Semibold_18,
                       ),
                     ),
                   ),
-                ],
-              ),
+              ],
             ),
+          ),
 
-            /// Expense
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20),
-              child: Row(
-                children: [
-                  const SizedBox(width: 10),
-                  Text(
-                    "지출",
-                    style: customTextStyle(
-                      fontFamily: Pretendard_Semibold_16,
-                      color: neutral_400,
-                    ),
-                  ),
-                  const Spacer(),
-                  Text(
-                    "${formatAmount(totalExpense)} 원",
-                    style: customTextStyle(
-                      fontFamily: Pretendard_Medium_24,
-                      color: Colors.black,
-                    ),
-                  ),
-                  const SizedBox(width: 10),
-                ],
-              ),
+          const SizedBox(height: 30),
+
+          /// Date Description
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 20),
+            child: Text(
+              _dateDisplay,
+              style: TextStyle(color: Colors.grey[600]),
             ),
+          ),
 
-            /// Income
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20),
-              child: Row(
-                children: [
-                  const SizedBox(width: 10),
-                  Text(
-                    "수입",
-                    style: customTextStyle(
-                      fontFamily: Pretendard_Semibold_16,
-                      color: neutral_400,
-                    ),
-                  ),
-                  const Spacer(),
-                  Text(
-                    "${formatAmount(totalIncome)} 원",
-                    style: customTextStyle(
-                      fontFamily: Pretendard_Medium_24,
-                      color: Colors.black,
-                    ),
-                  ),
-                  const SizedBox(width: 10),
-                ],
-              ),
-            ),
+          const Padding(
+            padding: EdgeInsets.symmetric(horizontal: 20, vertical: 10),
+            child: Divider(),
+          ),
 
-            const SizedBox(height: 20),
-
-            /// View Type Buttons
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 5),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                children: [
-                  for (var viewType in ['Daily', 'Weekly', 'Monthly'])
-                    ElevatedButton(
-                      onPressed: () => changeView(viewType),
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor:
-                            selectedView == viewType
-                                ? primary_400
-                                : primary_050,
-                        foregroundColor:
-                            selectedView == viewType
-                                ? Colors.white
-                                : primary_400,
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(6),
-                          side: const BorderSide(color: Colors.blue),
+          /// Combined List View
+          Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 20),
+            child:
+                (_incomeStats.isEmpty && _expenseStats.isEmpty)
+                    ? EmptyDiaryScreen()
+                    : Column(
+                      children: [
+                        ..._expenseStats.map(
+                          (item) => DailyCategoryItem(item: item),
                         ),
-                      ),
-                      child: Text(
-                        viewType,
-                        style: customTextStyle(
-                          fontFamily: Pretendard_Semibold_18,
+                        ..._incomeStats.map(
+                          (item) => DailyCategoryItem(item: item),
                         ),
-                      ),
+                      ],
                     ),
-                ],
-              ),
-            ),
+          ),
 
-            const SizedBox(height: 30),
-
-            /// Date Description
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20),
-              child: Text(
-                _dateDisplay,
-                style: TextStyle(color: Colors.grey[600]),
-              ),
-            ),
-
-            const Padding(
-              padding: EdgeInsets.symmetric(horizontal: 20, vertical: 10),
-              child: Divider(),
-            ),
-
-            /// Combined List View
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20),
-              child: Column(
-                children: [
-                  ..._expenseStats.map((item) => DailyCategoryItem(item: item)),
-                  ..._incomeStats.map((item) => DailyCategoryItem(item: item)),
-                ],
-              ),
-            ),
-
-            const SizedBox(height: 20),
-          ],
-        ),
+          const SizedBox(height: 20),
+        ],
       ),
     );
   }
